pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.1",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "hypothesis>=6.131.18",
    "opentelemetry-exporter-otlp-proto-grpc>=1.33.1",
    "opentelemetry-instrumentation-fastapi>=0.54b1",
    "opentelemetry-sdk>=1.33.1",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-benchmark>=5.1.0",
    "redis>=6.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "xgboost>=3.0.1",
]
